Move testutil/utils code into simply testutil. Repetitive name
