+++ tests/python/pants_test/BUILD
@@ -10,9 +10,9 @@ python_library(
     'tests/python/pants_test/jvm:nailgun_task_test_base',
     'tests/python/pants_test/jvm:jvm_tool_task_test_base',
     'tests/python/pants_test/engine:engine_test_base',
+    'tests/python/pants_test/engine:util',
     'tests/python/pants_test/subsystem:subsystem_utils',
     'tests/python/pants_test/base:context_utils',
-    'tests/python/pants_test/engine:util',
     'tests/python/pants_test/option/util',
     'tests/python/pants_test/testutils:file_test_util',
   ],
