+++ tests/python/pants_test/testutils/BUILD
@@ -5,7 +5,7 @@ python_library(
   name = 'mock_logger',
   sources = globs('mock_logger.py'),
   dependencies = [
-    'src/python/pants/testutil/utils:mock_logger',
+    'src/python/pants/testutil:mock_logger',
   ],
 )
 
@@ -15,7 +15,7 @@ python_library(
     'file_test_util.py',
   ],
   dependencies = [
-    'src/python/pants/testutil/utils:file_test_util',
+    'src/python/pants/testutil:file_test_util',
   ],
 )
 
@@ -25,7 +25,7 @@ python_library(
     'git_util.py',
   ],
   dependencies = [
-    'src/python/pants/testutil/utils:git_util',
+    'src/python/pants/testutil:git_util',
   ],
 )
 
@@ -35,7 +35,7 @@ python_library(
     'process_test_util.py',
   ],
   dependencies = [
-    'src/python/pants/testutil/utils:process_test_util',
+    'src/python/pants/testutil:process_test_util',
   ],
 )
 
@@ -45,6 +45,6 @@ python_library(
     'pexrc_util.py',
   ],
   dependencies = [
-    'src/python/pants/testutil/utils:pexrc_util',
+    'src/python/pants/testutil:pexrc_util',
   ],
 )
