+++ tests/python/pants_test/testutils/git_util.py
@@ -1,6 +1,6 @@
 # Copyright 2016 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testutil.utils.git_util import MIN_REQUIRED_GIT_VERSION as MIN_REQUIRED_GIT_VERSION  # noqa
-from pants.testutil.utils.git_util import git_version as git_version  # noqa
-from pants.testutil.utils.git_util import initialize_repo as initialize_repo  # noqa
+from pants.testutil.git_util import MIN_REQUIRED_GIT_VERSION as MIN_REQUIRED_GIT_VERSION  # noqa
+from pants.testutil.git_util import git_version as git_version  # noqa
+from pants.testutil.git_util import initialize_repo as initialize_repo  # noqa
