+++ tests/python/pants_test/testutils/pexrc_util.py
@@ -1,6 +1,6 @@
 # Copyright 2017 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testutil.utils.pexrc_util import (
+from pants.testutil.pexrc_util import (
   setup_pexrc_with_pex_python_path as setup_pexrc_with_pex_python_path,
 )  # noqa
