+++ tests/python/pants_test/testutils/process_test_util.py
@@ -1,14 +1,12 @@
 # Copyright 2017 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testutil.utils.process_test_util import ProcessStillRunning as ProcessStillRunning  # noqa
-from pants.testutil.utils.process_test_util import (
-  TrackedProcessesContext as TrackedProcessesContext,
-)  # noqa
-from pants.testutil.utils.process_test_util import _make_process_table as _make_process_table  # noqa
-from pants.testutil.utils.process_test_util import (
+from pants.testutil.process_test_util import ProcessStillRunning as ProcessStillRunning  # noqa
+from pants.testutil.process_test_util import TrackedProcessesContext as TrackedProcessesContext  # noqa
+from pants.testutil.process_test_util import _make_process_table as _make_process_table  # noqa
+from pants.testutil.process_test_util import (
   _safe_iter_matching_processes as _safe_iter_matching_processes,
 )  # noqa
-from pants.testutil.utils.process_test_util import (
+from pants.testutil.process_test_util import (
   no_lingering_process_by_command as no_lingering_process_by_command,
 )  # noqa
